+++ include/util/StringUtil.h
@@ -17,14 +17,17 @@ limitations under the License.
 */
 #pragma once
 
-#include <wx/wx.h>
+#include <wx/string.h>  // for wxString
+
+#include <cstdint>  // for int64_t
 
 namespace cszb_scoreboard {
 
 class StringUtil {
  public:
   static auto intToString(int value) -> wxString;
-  static auto stringToInt(const wxString &string, int default_value = 0) -> int32_t;
+  static auto stringToInt(const wxString &string, int default_value = 0)
+      -> int64_t;
 };
 
 }  // namespace cszb_scoreboard
